LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-animation-expected.txt


PASS @keyframes works with @property
PASS @keyframes picks up the latest @property in the document
FAIL Ongoing animation picks up redeclared custom property assert_equals: expected "rgb(150, 150, 150)" but got "0px"
PASS Ongoing animation matches new keyframes against the current registration
FAIL Ongoing animation picks up redeclared intial value assert_equals: expected "250px" but got "200px"
FAIL Ongoing animation picks up redeclared inherits flag assert_equals: expected "250px" but got "200px"
FAIL Ongoing animation picks up redeclared meaning of 'unset' assert_equals: expected "250px" but got "200px"
FAIL Transitioning from initial value assert_equals: expected "rgb(128, 0, 128)" but got "rgb(0, 0, 255)"
FAIL Transitioning from specified value assert_equals: expected "rgb(0, 64, 128)" but got "rgb(0, 128, 0)"
FAIL Transition triggered by initial value change assert_equals: expected "150px" but got "200px"
PASS No transition when changing types
FAIL No transition when removing @property rule assert_equals: expected " 100px" but got "100px"
